refactor: more moving and cleaning up node exports
