+++ .changeset/plugins-hmr-import.md
@@ -2,4 +2,4 @@
 '@seedcord/plugins': patch
 ---
 
-Import `HmrModuleHandler` from `@seedcord/core/hmr`. Adds `@seedcord/core` as a runtime dependency.
+`@seedcord/core` is now a runtime dependency.
