--- .changeset/seedcord-hmr-import.md
@@ -1,5 +1,0 @@
----
-seedcord: patch
----
-
-Import `wrapHot` from `@seedcord/core/hmr`.
