+++ .changeset/types-hmr-public.md
@@ -1,0 +1,5 @@
+---
+'@seedcord/types': minor
+---
+
+Export `HmrEventType`, `HmrUpdateEvent`, and `HmrAware` from the package root, for typing `onHmr` implementations.
