+++ packages/core/src/hmr.index.ts
@@ -1,2 +1,1 @@
-export { HmrManager } from '@hmr/HmrManager';
 export { HmrModuleHandler } from '@hmr/HmrModuleHandler';
