+++ packages/core/src/hmr/HmrModuleHandler.ts
@@ -6,7 +6,7 @@ import { formatFilePath } from '@seedcord/utils';
 import chalk from 'chalk';
 
 import type { Logger } from '@seedcord/logger';
-import type { HmrUpdateEvent } from '@seedcord/types/internal';
+import type { HmrUpdateEvent } from '@seedcord/types';
 
 interface HmrStore<THandler, TMiddleware, TArtifacts> {
     fileToHandlers: Map<string, Set<THandler>>;
