+++ packages/docs-engine/tsconfig.json
@@ -14,6 +14,6 @@
             "@remote/*": ["./src/remote/*"]
         }
     },
-    "include": ["src/**/*.ts", "tests/**/*.ts", "tsup.config.ts", "vitest.config.ts"],
+    "include": ["src/**/*.ts", "tests/**/*.ts", "tsdown.config.ts", "vitest.config.ts"],
     "exclude": ["node_modules", "dist", "generated"]
 }
