+++ packages/docs-generator/tsconfig.json
@@ -4,6 +4,6 @@
     "compilerOptions": {
         "rootDir": "."
     },
-    "include": ["src/**/*.ts", "tests/**/*.ts", "tsup.config.ts", "vitest.config.ts"],
+    "include": ["src/**/*.ts", "tests/**/*.ts", "tsdown.config.ts", "vitest.config.ts"],
     "exclude": ["node_modules", "dist", "generated"]
 }
