+++ packages/gateway/src/bot/controllers/CommandRegistry.ts
@@ -7,7 +7,8 @@ import { CommandMetadataKey, getDevChannel } from '@seedcord/core/internal';
 import { SeedcordErrorCode } from '@seedcord/errors';
 import { SeedcordError } from '@seedcord/errors/internal';
 import { Logger, paint } from '@seedcord/logger';
-import { formatFilePath, traverseDirectory } from '@seedcord/utils';
+import { formatFilePath } from '@seedcord/utils';
+import { traverseDirectory } from '@seedcord/utils/node';
 import chalk from 'chalk';
 import { Collection } from 'discord.js';
 import { Envapter } from 'envapt';
@@ -20,7 +21,7 @@ import type { ContextMenuCommandBuilder } from '@discordjs/builders';
 import type { Core } from '@interfaces/Core';
 import type { Initializeable } from '@interfaces/Plugin';
 import type { CommandMeta } from '@seedcord/core/internal';
-import type { HmrAware, HmrUpdateEvent } from '@seedcord/types/internal';
+import type { HmrAware, HmrUpdateEvent } from '@seedcord/types';
 import type { ApplicationCommand, Snowflake } from 'discord.js';
 
 type CommandCtor = new () => BuilderComponent<'command' | 'context_menu'>;
