+++ packages/gateway/src/bot/controllers/InteractionDispatcher.ts
@@ -13,7 +13,8 @@ import {
 import { SeedcordErrorCode } from '@seedcord/errors';
 import { SeedcordError } from '@seedcord/errors/internal';
 import { Logger, paint } from '@seedcord/logger';
-import { formatFilePath, hasKeys, traverseDirectory } from '@seedcord/utils';
+import { formatFilePath, hasKeys } from '@seedcord/utils';
+import { traverseDirectory } from '@seedcord/utils/node';
 import chalk from 'chalk';
 import { Events } from 'discord.js';
 import { Envapter } from 'envapt';
@@ -34,8 +35,7 @@ import type { Repliables, ValidInteractionTypes } from '@handlers/BaseHandler';
 import type { HandlerConstructor, InteractionMiddlewareConstructor } from '@handlers/constructors';
 import type { Core } from '@interfaces/Core';
 import type { Initializeable } from '@interfaces/Plugin';
-import type { CustomIdMatcher } from '@seedcord/types';
-import type { HmrAware, HmrUpdateEvent } from '@seedcord/types/internal';
+import type { CustomIdMatcher, HmrAware, HmrUpdateEvent } from '@seedcord/types';
 import type {
     AutocompleteInteraction,
     ButtonInteraction,
