+++ packages/plugins/tsconfig.json
@@ -8,6 +8,6 @@
             "@src/*": ["./src/*"]
         }
     },
-    "include": ["src/**/*.ts", "tests/**/*.ts", "tsup.config.ts", "vitest.config.ts"],
+    "include": ["src/**/*.ts", "tests/**/*.ts", "tsdown.config.ts", "vitest.config.ts"],
     "exclude": ["node_modules", "dist"]
 }
