+++ packages/seedcord/src/commands/dev/runtime/HmrPlugin.ts
@@ -8,13 +8,8 @@ import { minimatch } from 'minimatch';
 
 import type { DevEvent } from './events';
 import type { ResolvedSeedcordDevConfig } from '@core/config/schema';
-import type {
-    DevChannel,
-    HmrEventType,
-    HmrUpdateEvent,
-    SeedcordCliEvents,
-    SeedcordFrameworkEvents
-} from '@seedcord/types/internal';
+import type { HmrEventType, HmrUpdateEvent } from '@seedcord/types';
+import type { DevChannel, SeedcordCliEvents, SeedcordFrameworkEvents } from '@seedcord/types/internal';
 import type {
     EnvironmentModuleNode,
     HotUpdateOptions,
