+++ packages/seedcord/src/index.ts
@@ -6,5 +6,5 @@ export {
     type SeedcordHmrConfig
 } from '@core/config/schema';
 
-/** Package version. Falls back to '0.0.0' when run unbuilt (dev); the real value is injected at build time. */
+/** Package version */
 export const version = process.env.PACKAGE_VERSION ?? '0.0.0';
