+++ packages/seedcord/tsconfig.json
@@ -21,7 +21,7 @@
         "tests/**/*.tsx",
         "playground/**/*.ts",
         "playground/**/*.tsx",
-        "tsup.config.ts",
+        "tsdown.config.ts",
         "vitest.config.ts"
     ],
     "exclude": ["node_modules", "dist"]
