+++ packages/types/src/Types/index.ts
@@ -1,6 +1,7 @@
 export type * from './Colors';
 export type * from './Epoch';
 export type * from './Filters';
+export type * from './Hmr';
 export type * from './ObjectManipulation';
 export type * from './Misc';
 export type * from './TupleArrays';
