+++ packages/types/tsconfig.json
@@ -5,6 +5,6 @@
         "rootDir": "./",
         "outDir": "./dist"
     },
-    "include": ["src/**/*.ts", "tests/**/*.ts", "tsup.config.ts"],
+    "include": ["src/**/*.ts", "tests/**/*.ts", "tsdown.config.ts"],
     "exclude": ["node_modules", "dist"]
 }
