+++ packages/utils/src/node/directory.ts
@@ -63,44 +63,3 @@ export async function traverseDirectory(
         }
     }
 }
-
-/**
- * Options for formatting file paths.
- */
-export interface FormatFileOptions {
-    /**
-     * Whether to return only the directory part of the path.
-     *
-     * @defaultValue `false`
-     */
-    onlyDir?: boolean;
-    /**
-     * A prefix to prepend to the formatted path.
-     *
-     * @defaultValue `'./'`
-     */
-    prefix?: string;
-}
-
-/**
- * Formats a file path relative to the current working directory.
- * @param filePath - The file path to format.
- * @param options - Formatting options.
- * @returns The formatted file path.
- *
- * @example
- * ```ts
- * // cwd is /repo
- * formatFilePath('/repo/src/Bot.ts'); // './src/Bot.ts'
- * formatFilePath('/repo/src/Bot.ts', { onlyDir: true }); // './src'
- * formatFilePath('/repo/src/Bot.ts', { prefix: '' }); // 'src/Bot.ts'
- * ```
- */
-export function formatFilePath(filePath: string, options: FormatFileOptions = {}): string {
-    const { onlyDir = false, prefix = './' } = options;
-
-    const resolved = onlyDir
-        ? path.relative(process.cwd(), filePath.replace(/\/[^/]*$/, ''))
-        : path.relative(process.cwd(), filePath);
-    return `${prefix}${resolved}`;
-}
