+++ packages/utils/tsdown.config.ts
@@ -1,3 +1,3 @@
 import { createTsdownConfig } from '@seedcord/tsdown-config';
 
-export default createTsdownConfig({ entry: ['src/index.ts', 'src/internal.index.ts'] });
+export default createTsdownConfig({ entry: ['src/index.ts', 'src/internal.index.ts', 'src/node.ts'] });
